+++ packages/paragraph/src/index.ts
@@ -1,9 +1,17 @@
-import { type IActionPlugin } from '@rabbitholegg/questdk'
+import {
+  type ActionParams,
+  type IActionPlugin,
+  type MintActionParams,
+} from '@rabbitholegg/questdk-plugin-utils'
 
 import {
-  mint,
+  getFees,
+  getMintIntent,
+  getProjectFees,
   getSupportedChainIds,
   getSupportedTokenAddresses,
+  mint,
+  simulateMint,
 } from './Paragraph.js'
 
 export const Paragraph: IActionPlugin = {
@@ -11,4 +19,10 @@ export const Paragraph: IActionPlugin = {
   getSupportedTokenAddresses,
   getSupportedChainIds,
   mint,
+  getProjectFees: async (params: ActionParams) =>
+    getProjectFees(params as unknown as MintActionParams),
+  getFees: async (params: ActionParams) =>
+    getFees(params as unknown as MintActionParams),
+  getMintIntent,
+  simulateMint,
 }