+++ packages/registry/src/index.ts
@@ -36,6 +36,8 @@ import { Manifold } from '@rabbitholegg/questdk-plugin-manifold'
 import { Fabric } from '@rabbitholegg/questdk-plugin-fabric'
 import { Paragraph } from '@rabbitholegg/questdk-plugin-paragraph'
 import { Aerodrome } from '@rabbitholegg/questdk-plugin-aerodrome'
+import { Pods } from '@rabbitholegg/questdk-plugin-pods'
+import { Kwenta } from '@rabbitholegg/questdk-plugin-kwenta'
 import { BaseBridge } from '@rabbitholegg/questdk-plugin-basebridge'
 import { ENTRYPOINT } from './contract-addresses'
 import {
@@ -96,6 +98,8 @@ export const plugins: Record<string, IActionPlugin> = {
   [Fabric.pluginId]: Fabric,
   [Paragraph.pluginId]: Paragraph,
   [Aerodrome.pluginId]: Aerodrome,
+  [Pods.pluginId]: Pods,
+  [Kwenta.pluginId]: Kwenta,
   [BaseBridge.pluginId]: BaseBridge,
 }