+++ packages/artblocks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rabbitholegg/questdk-plugin-artblocks
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [[`59ae9b8d`](https://github.com/rabbitholegg/questdk-plugins/commit/59ae9b8df08fe4e05f7743f67a685428ea0927e6), [`cd6215e7`](https://github.com/rabbitholegg/questdk-plugins/commit/cd6215e78e6d42bc6d2e23fb46b7906c990f7c87)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.17
+
 ## 1.0.0-alpha.2
 
 ### Minor Changes